Reading images, datasets, bias and sigmoid
It's now possible to read brightness values from images. This has been implemented using the stb_image.h library.
Generated and modified a temporal python script using ChatGPT to produce squares and circles of different sizes and positions, to be used as a dataset to train the neural networks. This is just temporary, to be able to do quick tests.
Implemented the sigmoid activation function in neurons, so they don't have a linear behavior.
I'm currently working on adding bias to the neurons.