Big optimizations
I made big optimizations regarding mutated networks being written to disk. Now, only when there is a difference in performance, that network biases and weights are written to disk. This saves a lot of disk space and makes the genetic algorithm much faster.
Multithreading in training has also been implemented, which greatly increases the algorithm performance. The number of threads can be specified when creating the GeneticNetworkTrainer object.