Modified bindings and gui to adapt to the network changes
bindings.cpp and gui.py now correctly work after the changes made to many of the network trainer's functions. It's not possible to train a network to differentiate between more than 2 objects yet, through the gui.
It's now necessary to specify 2 neurons in the output layers for binary classification, instead of 1.