Fix incorrect linking of objects to image folders
Fixed a critical bug that made the networks not learn correctly what the objects were. The image path sent to the trainer didn't have the same order as the object path.
Also fixed a bug where to see the changes on a trained model, you had to restart the GUI.